fix bug with list shrink scrolling
