+++ src/components/wallet/sidebarWallet/SidebarWallet.tsx
@@ -15,7 +15,7 @@ import {
     BuyTariButton,
     DetailsCardBottomContent,
 } from './styles.ts';
-import { useCallback, useRef, useState, useEffect } from 'react';
+import { useCallback, useRef, useState } from 'react';
 import { HistoryListWrapper } from '@app/components/wallet/components/history/styles.ts';
 import { List } from '@app/components/transactions/history/List.tsx';
 import { open } from '@tauri-apps/plugin-shell';
@@ -41,16 +41,8 @@ export default function SidebarWallet({ section, setSection }: SidebarWalletProp
     const { data: xcData } = useFetchExchangeBranding();
     const detailsItem = useWalletStore((s) => s.detailsItem);
 
-    const targetRef = useRef<HTMLDivElement>(null);
+    const targetRef = useRef<HTMLDivElement>(null) as React.RefObject<HTMLDivElement>;
     const [isScrolled, setIsScrolled] = useState(false);
-
-    useEffect(() => {
-        const el = targetRef.current;
-        if (!el) return;
-        const onScroll = () => setIsScrolled(el.scrollTop > 1);
-        el.addEventListener('scroll', onScroll);
-        return () => el.removeEventListener('scroll', onScroll);
-    }, []);
 
     const isSyncing = useWalletStore((s) => s.wallet_scanning.is_scanning);
     const isSwapping = useWalletStore((s) => s.is_swapping);
@@ -103,7 +95,7 @@ export default function SidebarWallet({ section, setSection }: SidebarWalletProp
                     </AnimatePresence>
 
                     <HistoryListWrapper ref={targetRef}>
-                        <List />
+                        <List setIsScrolled={setIsScrolled} targetRef={targetRef} />
                     </HistoryListWrapper>
                 </>
             )}
